requirements.txt
google-genai
requests
beautifulsoup4
pydantic
